Added in one more tokenizer_utils file and norminetted 1 & 3.
Updated makefile and minishell.h for it too
