+++ src/tokenizer/tokenizer_utils_2.c
@@ -5,60 +5,97 @@
 /*                                                    +:+ +:+         +:+     */
 /*   By: axlee <axlee@student.42.fr>                +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
-/*   Created: 2024/04/19 12:33:25 by gyong-si          #+#    #+#             */
-/*   Updated: 2024/06/10 12:38:57 by axlee            ###   ########.fr       */
+/*   Created: 2024/06/10 12:13:57 by axlee             #+#    #+#             */
+/*   Updated: 2024/06/10 13:01:45 by axlee            ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "minishell.h"
 
-// check if character is space
-int	ft_iswhitespace(char *line)
+// print the linked list that holds the tokens;
+void	print_tokenlst(t_token *token_lst)
 {
-	if (*line == ' ')
-		return (1);
-	return (0);
-}
+	t_token	*curr;
 
-int	ft_isbackslash(char *line)
-{
-	if (*line == '\\')
-		return (1);
-	return (0);
-}
-
-int	ft_issemicolon(char *line)
-{
-	if (*line == ';')
-		return (1);
-	return (0);
-}
-
-char	*ft_copy(char *dest, const char *src, size_t n)
-{
-	size_t	i;
-
-	i = 0;
-	while (i < n && src[i] != '\0')
+	curr = token_lst;
+	while (curr != NULL)
 	{
-		dest[i] = src[i];
-		i++;
+		printf("token: %s\n", curr->token);
+		printf("token type: %d\n", curr->type);
+		curr = curr->next;
 	}
-	dest[i] = '\0';
-	return (dest);
 }
 
-char	*ft_strcat(char *dest, const char *src)
+void	handle_environment_variable(char **line, t_token **token_lst,
+		t_shell *minishell)
 {
-	char	*ptr;
+	char	*start;
+	char	*var_name;
+	char	*expanded;
+	char	*result;
+	int		var_len;
 
-	ptr = dest;
-	while (*ptr)
-		ptr++;
-	while (*src)
+	start = *line;
+	// Move past the dollar sign
+	(*line)++;
+	start = *line;
+	// Special case for $?
+	if (**line == '?')
 	{
-		*ptr++ = *src++;
+		(*line)++;
+		expanded = ft_itoa(minishell->last_return);
+		// Convert last return value to string
+		var_len = strlen(expanded);
+		result = malloc(var_len + 1);
+		if (result)
+		{
+			strcpy(result, expanded);
+			free(expanded);
+			// Handle the rest of the line as part of the same token
+			while (**line && !ft_iswhitespace(*line) && **line != '$'
+				&& **line != '\'' && **line != '\"')
+			{
+				result = ft_realloc(result, strlen(result) + 2);
+				strncat(result, *line, 1);
+				(*line)++;
+			}
+			token_add_back(token_lst, result, T_IDENTIFIER);
+			free(result);
+		}
+		return ;
 	}
-	*ptr = '\0';
-	return (dest);
+	// Find the end of the variable name
+	while (**line && (ft_isalnum(**line) || **line == '_'))
+		(*line)++;
+	var_len = *line - start;
+	var_name = ft_strndup(start, var_len);
+	expanded = get_env_value(minishell, var_name);
+	free(var_name);
+	if (expanded)
+	{
+		// Allocate space for the result
+		result = malloc(strlen(expanded) + 1);
+		if (result)
+		{
+			strcpy(result, expanded);
+			free(expanded);
+			// Handle the rest of the line as part of the same token
+			while (**line && !ft_iswhitespace(*line) && **line != '$'
+				&& **line != '\'' && **line != '\"')
+			{
+				result = ft_realloc(result, strlen(result) + 2);
+				strncat(result, *line, 1);
+				(*line)++;
+			}
+			token_add_back(token_lst, result, T_IDENTIFIER);
+			free(result);
+		}
+	}
+	else
+	{
+		// If the variable is not found, add the original text
+		result = ft_strndup(start - 1, var_len + 1);
+		token_add_back(token_lst, result, T_IDENTIFIER);
+		free(result);
+	}
 }
